Move app.js to components directory
