+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './app';
+import App from './components/app';
 import './index.css';
 import configureStore from './store/configureStore'
 import rootSaga from './sagas'
